Customer Segmentation sample directly using API instead of common-code with ModelBuilder, etc.
