--- samples/csharp/getting-started/Clustering_CustomerSegmentation/CustomerSegmentation.Train/ModelHelpers.cs
@@ -1,78 +1,0 @@
-﻿using Microsoft.ML.Runtime.Data;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-
-namespace CustomerSegmentation
-{
-    public static class ModelHelpers
-    {
-        static FileInfo _dataRoot = new FileInfo(typeof(Program).Assembly.Location);
-
-        public static string GetAssetsPath(params string[] paths)
-        {
-            if (paths == null || paths.Length == 0)
-                return null;
-
-            return Path.Combine(paths.Prepend(_dataRoot.Directory.FullName).ToArray());
-        }
-
-        public static string DeleteAssets(params string[] paths)
-        {
-            var location = GetAssetsPath(paths);
-
-            if (!string.IsNullOrWhiteSpace(location) && File.Exists(location))
-                File.Delete(location);
-            return location;
-        }
-
-        public static IEnumerable<string> Columns<T>() where T : class
-        {
-            return typeof(T).GetProperties().Select(p => p.Name);
-        }
-
-        public static IEnumerable<string> Columns<T, U>() where T : class
-        {
-            var typeofU = typeof(U);
-            return typeof(T).GetProperties().Where(c => c.PropertyType == typeofU).Select(p => p.Name);
-        }
-
-        public static IEnumerable<string> Columns<T, U, V>() where T : class
-        {
-            var typeofUV = new[] { typeof(U), typeof(V) };
-            return typeof(T).GetProperties().Where(c => typeofUV.Contains(c.PropertyType)).Select(p => p.Name);
-        }
-
-        public static IEnumerable<string> Columns<T, U, V, W>() where T : class
-        {
-            var typeofUVW = new[] { typeof(U), typeof(V), typeof(W) };
-            return typeof(T).GetProperties().Where(c => typeofUVW.Contains(c.PropertyType)).Select(p => p.Name);
-        }
-
-        public static string[] ColumnsNumerical<T>() where T : class
-        {
-            return Columns<T, float, int>().ToArray();
-        }
-
-        public static string[] ColumnsString<T>() where T : class
-        {
-            return Columns<T, string>().ToArray();
-        }
-
-        public static string[] ColumnsDateTime<T>() where T : class
-        {
-            return Columns<T, DateTime>().ToArray();
-        }
-
-        //public static IEnumerable<string> GetColumnNames(this ISchema schema)
-        //{
-        //    for (int i = 0; i < schema.ColumnCount; i++)
-        //    {
-        //        if (!schema.IsHidden(i))
-        //            yield return schema.GetColumnName(i);
-        //    }
-        //}
-    }
-
-}
